hotfixing Jindra s push, to be tested but only openid login might be affected, everything else should work

+++ src/gui/src/main/java/cz/filmtit/client/FilmTitServiceHandler.java
@@ -110,7 +110,7 @@ public class FilmTitServiceHandler {
 	
 
 	    
-	public void validateAuthentication (String responseURL, long authID, AuthenticationValidationWindow authenticationValidationWindow) {
+	public void validateAuthentication (String responseURL, int authID, AuthenticationValidationWindow authenticationValidationWindow) {
 		new ValidateAuthentication(responseURL, authID, authenticationValidationWindow);
 	}
